backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
ultralytics==8.3.50
python-multipart==0.0.20
Pillow==11.1.0
python-socketio==5.12.1
